# docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/recipe/{id}": {
      "get": {
        "tags": [
          "Recipe controller"
        ],
        "summary": "Get a single recipe by ID",
        "operationId": "getRecipe",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Recipe not found"
          },
          "200": {
            "description": "Found the recipe",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Recipe"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Recipe controller"
        ],
        "summary": "Update a recipe",
        "operationId": "updateRecipe",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Recipe"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Recipe not found"
          },
          "200": {
            "description": "Recipe updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Recipe"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Recipe controller"
        ],
        "summary": "Delete a recipe",
        "operationId": "deleteRecipe",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not found"
          },
          "200": {
            "description": "Recipe deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Recipe"
                }
              }
            }
          }
        }
      }
    },
    "/recipe": {
      "get": {
        "tags": [
          "Recipe controller"
        ],
        "summary": "Filter the recipes",
        "operationId": "filterRecipe",
        "parameters": [
          {
            "name": "vegan",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "servings",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "ingredient",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "include",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "instructions",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Found the recipe",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Recipe"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Recipe controller"
        ],
        "summary": "Create a recipe",
        "operationId": "createRecipe",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Recipe"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Recipe created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Recipe"
                }
              }
            }
          },
          "400": {
            "description": "Wrong data"
          }
        }
      }
    },
    "/recipe/list": {
      "get": {
        "tags": [
          "Recipe controller"
        ],
        "summary": "List all the recipes",
        "operationId": "listRecipes",
        "responses": {
          "200": {
            "description": "Recipe list generated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Recipe"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Recipe": {
        "required": [
          "ingredients",
          "instructions",
          "name",
          "servings"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "servings": {
            "type": "integer",
            "format": "int64"
          },
          "ingredients": {
            "type": "string"
          },
          "instructions": {
            "type": "string"
          },
          "vegan": {
            "type": "boolean"
          }
        }
      }
    }
  }
}